requirements.txt
anthropic>=0.44.0
beautifulsoup4==4.14.3
Flask==3.1.3
playwright==1.58.0
pypdf==6.9.0
python-dotenv==1.2.1
requests==2.32.5
websockets==15.0.1
